requirements.txt
sentence-transformers
faiss-cpu
transformers
torch
pypdf
indic-transliteration
regex
numpy
flask